arm: dts: k3-am62p5-var-som-u-boot: enable dm i2c in SPL for all enabled bus
Without this changes when the uclass_get_device_by_seq search for the main_i2c3 bus
it returns ENODEV error. This means that the EEPROM I2C bus is not found.
Enabling only main_i2c3 the device name is associated to the i2c index 1 while the correct value
is 3.

Signed-off-by: FrancescoFerraro <francesco.f@variscite.com>